requirements.txt
boto==2.37.0
easy-thumbnails==2.2
django==1.7
django-js-reverse==0.5.0
django-storages-redux==1.2.3
django-suit==0.2.12
django-tastypie==0.12.1
gevent==1.0.1
greenlet==0.4.5
gunicorn==18.0
dj_database_url==0.3.0
django-cache-url==0.8.0
psycopg2==2.6
django_debug_toolbar==1.3.0
Pillow==2.8.1
raven==5.2.0
pytz==2015.2
django-watson==1.1.7
jsonfield==1.0.3